Deljenje dokumenata urađeno

+++ src/app/actions/ActNodeDelete.java
@@ -2,6 +2,7 @@ package app.actions;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
+import java.util.ArrayList;
 
 import javax.swing.*;
 
@@ -11,6 +12,7 @@ import app.models.page.Page;
 import app.models.project.Project;
 import app.models.workspace.Workspace;
 import app.views.MainFrame;
+import app.views.document.move_documents.MoveDocumentsDialog;
 
 
 public class ActNodeDelete extends GAbstractAction {
@@ -35,6 +37,39 @@ public class ActNodeDelete extends GAbstractAction {
         if (selectedNode instanceof Project) {
 
             Project project = (Project) selectedNode;
+
+            ArrayList<Document> documentsToBeDeleted = new ArrayList<>();
+            for (AbstractNode node : project.getChildren()) {
+                Document document = (Document) node;
+
+                int parentCount = 0;
+                for (Project parent : document.getParents()) {
+                    if (parent != null)
+                        parentCount += 1;
+                }
+
+                if (parentCount < 2)
+                    documentsToBeDeleted.add(document);
+            }
+
+            // If we have documents for moving,
+            // delegate that action to MoveDocuments dialog:
+            if (!documentsToBeDeleted.isEmpty()) {
+                if (MainFrame.getInstance().getWorkspace().getChildCount() < 2)
+                    ErrorHandler.showDocumentDeletedDialog();
+                else
+                    new MoveDocumentsDialog(project, documentsToBeDeleted);
+
+                // Project deletion will be handled by the dialog.
+                return;
+            }
+
+            // Delete this Project as a parent:
+            for (AbstractNode node : project.getChildren()) {
+                Document document = (Document) node;
+                document.removeParent(project);
+            }
+
             Workspace parent = (Workspace) project.getParent();
             parent.deleteProject(project);
